create01_seeder

+++ database/migrations/2021_12_18_230908_create_followers_table.php
@@ -14,7 +14,7 @@ class CreateFollowersTable extends Migration
     public function up()
     {
         Schema::create('followers', function (Blueprint $table) {
-            $table->unsignedInteger('following_id')->comment('フォローしているユーザID');
+            $table->unsignedBigInteger('following_id')->comment('フォローしているユーザID');
             $table->unsignedInteger('followed_id')->comment('フォローされているユーザID');
 
             $table->index('following_id');
